separated publishers
separated the publishers for easier testing
