+++ .gitignore
@@ -1,0 +1,3 @@
+
+app/influxpub/influxpub
+app/tmpub/tmpub
